flip sep_by{,_1} parameters...
In Haskell, it makes sense for the first to be the main parser and the
second to be the separator because Haskell supports making any function
an operator, and so the way you'd use it is "p `sepBy` s". Wanting to
specify the separator before the main subparser is also no problem,
since one can do "(`sepBy` s) p".

Ruby doesn't have any equivalent syntax though, and in light of that
limitation, we have to take into account that more often than not, the
separator is the smaller expression. Therefore, it's more beneficial to
flip them and have the separator stated before the other. That way we
can do:

    sep_by(lit(","),
      # very big, multiline parser
    )