remove on_catch...
Test is failing since code regarding failures has changed so much. We
don't even use it now. It was introduced with the ArithmeticParser
example because of the contrieved parsing methodology we were using. I
can't see myself using this method anymore, so I think it's better to
just remove it.