remove splicing operators in favor of Splicer class...
A boolean marker on the Tree nodes wasn't enough as one start would
splice to ends meant for another start. We needed unique markers to link
starts and ends. Instances of Splicer fulfill that role as markers
(their object_id is unique).

This also meant changing splice_to_ends! and renaming it to splice_to!,
adding a marker parameter.